+++ Macaw/macaw.h
@@ -4,14 +4,14 @@
 #include "main/core.h"
 
 namespace macaw {
-	extern double setting_height;
+	extern float setting_height;
 	extern int setting_octaves;
 
 	const char* version();
 
 	void seed_set(unsigned int);
 	void generate(float*, int, int);
-	void set_height(double);
+	void set_height(float);
 	void set_octaves(int);
 	float* _gen_white_noise(int, int);
 	float* _gen_smooth_noise(float*, int, int, int);
